feat : LottoFinalResult.java를 LottoFinanceStatsCalculator로 수정 후, 메소드 구현
