+++ src/main/java/controller/LottoResultCalculatorController.java
@@ -3,14 +3,13 @@ package controller;
 import constants.LottoSettingsConstants;
 import constants.ScriptConstants;
 import dto.LottoResultDto;
-import model.Lotto;
 import model.LottoBatch;
+import model.LottoFinanceStatsCalculator;
 import model.LottoResult;
 import model.WinCondition;
 import view.InputView;
 import view.OutputView;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -27,34 +26,18 @@ public class LottoResultCalculatorController {
     }
 
     public void calculate() {
+        LottoFinanceStatsCalculator lottoFinanceStatsCalculator = generateLottoFinanceStatsCalculator();
+        this.outputView.printAllStats(wrapLottoIntoDto(lottoFinanceStatsCalculator.getLottoResults()));
+        this.outputView.printReturnRatio(lottoFinanceStatsCalculator.getReturnRatio());
+    }
+
+    protected LottoFinanceStatsCalculator generateLottoFinanceStatsCalculator() {
         List<Integer> winningNumbers = this.inputView.getWinningNumbers();
-        int bonusNumber = this.inputView.getSingleIntegerFromUserAfterShowingAScript(ScriptConstants.INPUT_ENTER_WINNING_NUMBER_SCRIPT);
+        int bonusNumber = this.inputView.getSingleIntegerFromUserAfterShowingAScript(ScriptConstants.INPUT_ENTER_BONUS_NUMBER_SCRIPT);
+
         WinCondition winCondition = new WinCondition(winningNumbers, bonusNumber);
 
-        List<LottoResult> matchCountPerLotto = this.getMatchCountPerLotto(winCondition);
-        this.outputView.printAllStats(wrapLottoIntoDto(matchCountPerLotto));
-        this.outputView.printReturnRatio(getReturnRatio(winCondition));
-    }
-
-    protected double getReturnRatio(WinCondition winCondition) {
-        List<LottoResult> result = this.getMatchCountPerLotto(winCondition);
-
-        double earnResult = 0.0;
-        for (LottoResult lottoResult : result) {
-            earnResult += lottoResult.getReward();
-        }
-
-        return earnResult / (LottoSettingsConstants.LOTTO_PRICE * lottoBatch.getLottoCount());
-    }
-
-    protected List<LottoResult> getMatchCountPerLotto(WinCondition winCondition) {
-        List<LottoResult> result = new ArrayList<>();
-
-        for (Lotto lotto : lottoBatch.getAllLotto()) {
-            result.add(lotto.compareWithWinCondition(winCondition));
-        }
-
-        return result;
+        return new LottoFinanceStatsCalculator(this.lottoBatch, winCondition);
     }
 
     protected LottoResultDto wrapLottoIntoDto (List<LottoResult> lottoResults) {
